corregido y norminette ok

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: anruiz-d <anruiz-d@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/19 01:20:34 by anruiz-d          #+#    #+#             */
-/*   Updated: 2024/09/27 02:00:17 by anruiz-d         ###   ########.fr       */
+/*   Updated: 2024/10/06 14:08:20 by anruiz-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
